+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphand/cli
 
+## 1.4.0
+
+### Minor Changes
+
+- improve gdx integration, fixes & improvements
+
+### Patch Changes
+
+- Updated dependencies
+  - @graphand/client-module-auth@1.4.0
+  - @graphand/client@1.5.0
+
 ## 1.3.11
 
 ### Patch Changes