+++ packages/module-realtime/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphand/client-module-realtime
 
+## 1.3.0
+
+### Minor Changes
+
+- improve gdx integration, fixes & improvements
+
+### Patch Changes
+
+- Updated dependencies
+  - @graphand/client@1.5.0
+  - @graphand/core@1.7.0
+
 ## 1.2.8
 
 ### Patch Changes